refactor: header scroll module 분리
